Vorbereitung für Punktespeicherung

+++ app/model/html/ElementResult.java
@@ -33,7 +33,7 @@ public abstract class ElementResult {
     return "Element " + elementName + " wurde nicht gefunden.";
   }
   
-  public double getPoints() {
+  public int getPoints() {
     double attrsFound = attrs.stream().filter(attr -> attr.isFound()).collect(Collectors.counting()) / 2.;
     return (success != Success.NONE) ? 1 + attrsFound : 0;
   }
